Standardize documentation in removedevices.js
Make the header comment documentation a combination (concatenation) of
all of the package documentation files to avoid any ambiguity and limit
confusion about where to find documentation.

This redundancy is a bit annoying and it will likely lead to the copies
getting out of sync, but I think it is worth having so that users can
only download/share the .js file and still have all information.

Note:  Relicensed from ISC license to MIT.  Functionally the same, but
the MIT is more generally recognized and has clearer permissions grants.

Signed-off-by: Kevin Locke <klocke@digitalenginesoftware.com>